Update DoctrineParamConverter.php
This changes the behaviour to map the ParamConverter keys to the repository method's signature. It would then be possible to have the following mapping with a different signature in the repository:

    * "mapping": {"permalink": "permalink", "year": "year", "month": "month", "day": "day"}
    findForYmdLink($year, $month, $day, $permalink)

I favour the solution where the developer is responsible to make sure the keys are in the right order. This seems very magically to me.