Allow multiple @Security
The current behavior is not cool because it allows user to use only one
@Security tag per action OR per class. BUT Nothing prevents user to put
a tag on the class AND another one on the action. And it's even worst
because only the tag on the action will be used.

Because almost all annotations allow multiple instances (like @Route,
@Template, @Method) I think it's really convenient to allow multiple
@Security, and more over to fix a bug.

I already see theses PR;
* https://github.com/sensiolabs/SensioFrameworkExtraBundle/pull/368
* https://github.com/sensiolabs/SensioFrameworkExtraBundle/pull/370/files
But they are not perfect.