Make getIdentifier protected to allow override
Change the getIdentifier method in the Doctrine param converter from
private to protected. This allows overriding the method in classes
that extend the param converter. This is useful when you fiddle with
your ID’s in the URL, for example use Hashids.

With the method set to private you can’t override it.