Add Hashids ParamConverter
ParamConverter that can decode attributes hashed with Hashids and pass them on to the inner ParamConverter (e.g. DoctrineParamConverter).

As the inner ParamConverter is free to choose I did not include a service definition. To use it you’ll need to add a service definition and set a priority (higher than the inner).